+++ R/data-encyclopedists.R
@@ -5,12 +5,13 @@
 #' France in the second half of the 18th century. The data set is available in
 #' English (`encyclopedists`) and French (`encyclopedists_fr`).
 #'
-#' @format A tibble with 4 rows and 11 variables:
+#' @format A tibble with 4 rows and 12 variables:
 #' \describe{
 #'   \item{given_name,prénom}{authors' given names}
 #'   \item{family_name,nom}{authors' family names}
 #'   \item{email,courriel}{authors' email addresses}
 #'   \item{phone,téléphone}{authors' phone numbers}
+#'   \item{orcid}{authors' ORCID}
 #'   \item{affiliation}{authors' affiliations}
 #'   \item{role_n}{authors' roles written as nouns}
 #'   \item{role_v}{authors' roles written as active verbs}